backend part completed

+++ backend/controllers/goalController.js
@@ -1,20 +1,20 @@
 const asyncHandler = require('express-async-handler')
-const GoalsModel = require('../models/GoalsModel')
 
-const Goal = GoalsModel 
-
+const Goal = require('../models/GoalsModel') 
+const User = require('../models/UserModel')
 //@desc Get Goals
 //@route GET api/Goals
 //@access Private => After adding middleware 'auth'
 const getGoals =  asyncHandler( async (req, res) => {
-    const goals = await Goal.find()
+            // Protect Routes => { user: req.user.id }
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)  
     })
 
 const getGoal =  asyncHandler( async (req, res) => {
     // const goal = await Goal.findOne(req.params.id)
-    const goal = await Goal.findOne({ _id: req.params.id });
+    const goal = await Goal.findOne({ _id: req.user.id });
 
 
     res.status(200).json(goal)  
@@ -33,7 +33,9 @@ const setGoal =  asyncHandler (async (req, res) => {
         throw new Error("Please Add A Text Field")  
     }
         const goal = await Goal.create({
-            text: req.body.text
+            text: req.body.text,
+            // Protect Routes
+            user: req.user.id
         })
     res.status(200).json(goal)  
 })
@@ -49,6 +51,20 @@ const updateGoal = asyncHandler (async  (req, res) => {
         res.status(404)
         throw new Error('aie aie aie! Goal Not Found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User Not Authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true
     })
@@ -63,6 +79,25 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
       return res.status(404).json({ error: 'Goal not found' });
     }
+    
+    
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User Not Authorized')
+    }
+
+
+
+
     await Goal.deleteOne({ _id: req.params.id });
     return res.status(200).json({ id: req.params.id });
   });
